# Add standard deviation and 25% and 75% quantiles to `describe` :detailed