@bkamins 2020-09-30
Can you please update the NEWS.md entry (I think we already have something about `describe`).