@nalimilan 2020-09-30
Performance benchmarks are a bit weird, as calling `describe` without arguments is slower than calling it with the same arguments, both on master and on this PR. Also I've made it a bit faster to compute only the median, as before we always computed q25 and q75 even when not used. But overall the impact isn't that large, so I'd say it's mostly a design decision.

```julia
# master
julia> df = DataFrame(rand(10, 1_000_000));

julia> @btime describe(df);
  4.183 s (14999664 allocations: 1.24 GiB)

julia> @btime describe(df, :mean, :min, :median, :max, :nmissing, :eltype);
  3.557 s (14999720 allocations: 1.24 GiB)

julia> @btime describe(df, :mean, :std, :min, :median, :max, :nmissing, :eltype);
  3.826 s (19999734 allocations: 1.32 GiB)

julia> @btime describe(df, :mean, :std, :min, :q25, :median, :q75, :max, :nmissing, :eltype);
  5.041 s (19999765 allocations: 1.33 GiB)

# This PR
julia> df = DataFrame(rand(10, 1_000_000));

julia> @btime describe(df);
  5.356 s (20999682 allocations: 1.35 GiB)

julia> @btime describe(df, :mean, :min, :median, :max, :nmissing, :eltype);
  3.103 s (13999720 allocations: 1.21 GiB)

julia> @btime describe(df, :mean, :std, :min, :median, :max, :nmissing, :eltype);
  3.495 s (18999734 allocations: 1.29 GiB)

julia> @btime describe(df, :mean, :std, :min, :q25, :median, :q75, :max, :nmissing, :eltype);
  4.226 s (20999765 allocations: 1.35 GiB)
```