@tk3369 2020-09-30
I like std but generally don’t look at q25/q75 by default.  Performance is more important to me.