@pdeffebach 2020-09-30
This is fine. Stata keeps `:q25` and `:q75` behind the `detail` option, so its natural we do as well.