@bkamins 2020-11-07
@nalimilan - please also update manual, as we showe `describe` there.