@nalimilan 2020-11-08
Actually I'm not sure it's a good idea. Maybe a better solution would be to support `:detailed` as a shortcut for adding standard deviation and quartiles, and keep the default output relatively short.