@bkamins 2020-11-08
we already have `:all`. So `:detailed` would be "in the middle"?