@bkamins 2020-11-08
OK - so this is non-breaking then and can be post 0.22. I understand that you will make the changes? Thank you!