@pdeffebach 2020-11-13
Just looked at this and I don't think it's a good idea. 

Summary statistics with the new printing are sitting at 77 characters. I think the addition of `:q25` and `:q75` would make it too wide and harder to read.