@bkamins 2020-11-13
> Just looked at this and I don't think it's a good idea.

What do you refer to exactly?

To summary the conclusion from the discussion is:
1. leave the default as is.
2. add `:detailed` option `(similar to `:all`) that would add std, Q25 and Q75 to printed statistics

It seems from your comment that you would be OK with this approach.