@pdeffebach 2020-11-13
Yeah sorry, I am happy with `detail` (though hope Stata doesn't get mad a la google v oracle). Just not as default.