@nalimilan 2021-11-05
I've pushed a new version which supports `:detailed`. The only difference with `:all` is that the first and last value are not reported, but these are likely to be less interesting as often datasets are not ordered.